Following changes applied:
Caves now appear more frequently farther underground, making the surface less ugly with cave scarring.
Changed the lighting of block faces when the point light is turned off (hit L) to make it easier to see detail in caves
World is now 2 chunks high instead of 1, so that we can enjoy more caves.